backend/requirements.txt
Flask
flask-cors
pymysql
pyjwt